\name{mcRigor-package}
\alias{mcRigor-package}
\alias{mcRigor}
\docType{package}
\title{
\packageTitle{mcRigor}
}
\description{
\packageDescription{mcRigor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mcRigor}
\packageIndices{mcRigor}
Detection of dubiuos metacells and optimization of metacell partitioning.
}
\author{
\packageAuthor{mcRigor}

Maintainer: \packageMaintainer{mcRigor}
}
\keyword{metacell partitioning, scRNA-seq, sparsity, technical noise}

