% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRigor_function.R
\name{mcRigor_covariate}
\alias{mcRigor_covariate}
\title{To assist metacell object building in determining the metacell covariates (metadata)}
\usage{
mcRigor_covariate(
  sc_covariate,
  sc_membership,
  method = c("jaccard", "relative", "absolute")
)
}
\arguments{
\item{sc_covariate}{a vector (or dataframe) of single cell covariates}

\item{sc_membership}{a vector (or dataframe) of metacell memberships of single cells}

\item{method}{method to define the most abundant cell covariate within metacells. Available: "jaccard" (default), "relative", "absolute".
\itemize{
  \item jaccard - assign metacell to covariate with the maximum jaccard coefficient (recommended)
  \item relative - assign metacell to covariate with the maximum relative abundance (normalized by cluster size), may result in assignment of metacells to poorly represented (small) covariate due to normalization
  \item absolute - assign metacell to covariate with the maximum absolute abundance within metacell, may result in disappearance of poorly represented (small) clusters
}}
}
\value{
a vector of assigned metacell covariates
}
\description{
To assist metacell object building in determining the metacell covariates (metadata)
}
