% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRigor_function.R
\name{mcRigor_projection}
\alias{mcRigor_projection}
\title{To visualize a given metacell partitioning by projecting metacells onto the single cell embedding space.}
\usage{
mcRigor_projection(
  obj_singlecell,
  sc_membership = NULL,
  sc.reduction = c("umap", "tsne", "pca"),
  dims = c(1, 2),
  metric = "size",
  add_testres = F,
  test_stats = NULL,
  Thre = NULL,
  test_cutoff = 0.01,
  color_field = NULL,
  cpalette = c("#13678A", "#45C4B0", "#9AEBA3", "#BF847E", "#F2C12E", "#7E57C2",
    "#FACFCE", "#9E9D24", "#DAFDBA", "#86ABD4", "#42A5F5", "#546E7A", "#D4E157",
    "#76FF03", "#6D4C41", "#004D40", "#AB47BC", "#D81B60"),
  sc.alpha = 0.5,
  mc.alpha = 1,
  pt_size = 1,
  axis_lab = F,
  max_mcsize = 1000,
  continuous_metric = T,
  dub_mc.label = F,
  dub_mc_test.label = F,
  label_text = F
)
}
\arguments{
\item{obj_singlecell}{the Seurat object of single cells}

\item{sc_membership}{A named vector (or dataframe) of the metacell membership of single cells}

\item{sc.reduction}{The single cell reduction method to produce single cell embeddings}

\item{dims}{The dimensions to use in the single cell embeddings}

\item{metric}{The variable that determines the sizes of dots representing the metacells}

\item{add_testres}{A bool indicating whether to add the mcRigor results (dubious or trustworthy) as part of obj_metacell's metadata}

\item{test_stats}{If add_testres = True, this argument is needed. Usually should be TabMC, an output from previous steps.}

\item{Thre}{The threshold for dubious metacell detection. If not inputed, it will be computed based on test_stats.}

\item{test_cutoff}{The test size for dubious metacell detection testing}

\item{color_field}{The variable based on which to color the cells.}
}
\value{
a scatter plot of metacells projected to the single cell 2D embedding space.
}
\description{
To visualize a given metacell partitioning by projecting metacells onto the single cell embedding space.
}
