% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRigor_function.R
\name{mcRigor_purity}
\alias{mcRigor_purity}
\title{To assist metacell object building in compute the purity of metacells with respect to each covariate}
\usage{
mcRigor_purity(
  sc_covariate,
  sc_membership,
  method = c("max_proportion", "entropy")[1]
)
}
\arguments{
\item{sc_covariate}{vector of single cell covariates}

\item{sc_membership}{vector of assignment of single-cell data to metacells}

\item{method}{method to compute metacell purity.
\code{"max_proportion"} if the purity is defined as a proportion of the most abundant covariate (cell type) within super-cell or
\code{"entropy"} if the purity is defined as the Shanon entropy of the covariates metacell consists of.}
}
\value{
a vector of metacell purity, which is defined as:
- proportion of the most abundant covariate within metacell for \code{method = "max_proportion"} or
- Shanon entropy for \code{method = "entropy"}.
With 1 meaning that metacell consists of single cells from one covariate (reference assignment)
}
\description{
To assist metacell object building in compute the purity of metacells with respect to each covariate
}
