% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRigor_function.R
\name{mcRigor_test}
\alias{mcRigor_test}
\title{mcRigor_test}
\usage{
mcRigor_test(
  TabMC,
  test_cutoff = 0.01,
  thre_smooth = T,
  thre_bw = 1/6,
  D_bw = 10,
  optim_method = "tradeoff",
  dub_rate = 0.1,
  weight = 0.5,
  draw = T,
  palpha = 1,
  org_color = c("red", "orange", "yellow", "lightblue"),
  null_color = "#666666",
  pur_metric = NULL
)
}
\arguments{
\item{TabMC}{A dataframe containing the permutation results. Saved in the previous steps}

\item{test_cutoff}{The test size for dubious metacell detection testing}

\item{thre_smooth}{A boolean indicating whether to smooth the threshold function}

\item{thre_bw}{If thre_smooth is True, what is the bandwidth for smoothing}

\item{D_bw}{A boolean indicating whether to smooth the dubious rate with respect to metacell size}

\item{optim_method}{The method used for granularity level optimization. Default is trading off between sparsity and dubious rate}

\item{dub_rate}{If tradeoff is not used for optimization, what is highest acceptable dubious rate}

\item{weight}{The weight for dubious rate in the tradeoff.}

\item{draw}{A boolean indicating whether to visualize the mcRigor results}

\item{palpha}{Point alpha value for transparency in drawing.}

\item{org_color}{The colors indicating metacell purities or other interested factors}

\item{null_color}{The color for the null.}

\item{pur_metric}{Can be NULL or a metadata variable name, ex. cell type.}
}
\value{
A list containing the following fields: 
\item{optimized}{The optimization results, containing the optimal gamma and its corresponding Sore}
\item{scores}{A data frame containing the evaluation scores for each gamma}
\item{threshold}{The thresholds for dubious metacell detection}
\item{TabMC}{A dataframe containing the permutation results, elements to calculate the test statistics mcDiv and mcDiv null}
\item{test_plot}{Plots showing the mcRigor results}
\item{plot_list}{A list of separate plots}
}
\description{
A building block of the main function mcRigor. To detect dubious metacells and optimize metacell partitioning based on the output permutation results (TabMC)
}
