% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRigor_function.R
\name{mcRigor_threshold}
\alias{mcRigor_threshold}
\title{A building block of the main functions. To derive the thresholds for detecting dubious metacells based on the output permutation results (TabMC)}
\usage{
mcRigor_threshold(
  TabMC,
  test_cutoff = 0.01,
  thre_smooth = T,
  thre_bw = 1/6,
  draw = T,
  palpha = 1,
  org_color = c("red", "orange", "yellow", "lightblue"),
  null_color = "#666666",
  pur_metric = NULL
)
}
\arguments{
\item{TabMC}{A dataframe containing the permutation results. Saved in the previous steps}

\item{test_cutoff}{The test size for dubious metacell detection testing}

\item{thre_smooth}{A boolean indicating whether to smooth the threshold function}

\item{thre_bw}{If thre_smooth is True, what is the bandwidth for smoothing}

\item{draw}{A boolean indicating whether to visualize the mcRigor results}

\item{palpha}{Point alpha value for transparency in drawing.}

\item{org_color}{The colors indicating metacell purities or other interested factors}

\item{null_color}{The color for the null.}

\item{pur_metric}{Name of the covariate that we want to compute purity on. Can be NULL or a metadata variable name, ex. cell type.}
}
\value{
A list containing the following fields: 
\item{threshold}{The thresholds for dubious metacell detection}
\item{TabMC}{A dataframe containing the permutation results and the testing results given by mcRigor}
\item{test_plot}{The scatter plots demonstrating the mcDiv values and the obtained thresholds for dubious metacell detection}
\item{purity_plot}{A violin plot showing the purity distribution of the pur_metric covariate in dubious metacells and trustworthy metacells}
}
\description{
A building block of the main functions. To derive the thresholds for detecting dubious metacells based on the output permutation results (TabMC)
}
