% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRigor_function.R
\name{mcRigor_tradeoff}
\alias{mcRigor_tradeoff}
\title{A building block of the main functions. To evaluate each metacell partition and optimize metacell partitioning based on the output permutation results (TabMC) and thresholds (threshold)}
\usage{
mcRigor_tradeoff(
  TabMC,
  threshold = NULL,
  D_bw = 10,
  optim_method = c("tradeoff", "dub_rate_large", "dub_rate_small"),
  dub_rate = 0.1,
  weight = 0.5,
  draw = T
)
}
\arguments{
\item{TabMC}{A dataframe containing the permutation results. Saved in the previous steps}

\item{threshold}{A dataframe containing the dubious metacell detection thresholds given by mcRigor_threshold}

\item{D_bw}{A boolean indicating whether to smooth the dubious rate with respect to metacell size}

\item{optim_method}{The method used for granularity level optimization. Default is trading off between sparsity and dubious rate}

\item{dub_rate}{If tradeoff is not used for optimization, what is highest acceptable dubious rate}

\item{weight}{The weight for dubious rate in the tradeoff.}

\item{draw}{A boolean indicating whether to visualize the mcRigor results}
}
\value{
A list containing the following fields: 
\item{optimized}{The optimization results, containing the optimal gamma and its corresponding Sore}
\item{scores}{A data frame containing the evaluation scores for each gamma}
\item{optim_plot}{The line plot to visualize the tradeoff for hyperparameter opimization.}
}
\description{
A building block of the main functions. To evaluate each metacell partition and optimize metacell partitioning based on the output permutation results (TabMC) and thresholds (threshold)
}
