% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRigor_twostep.R
\name{mcRigorTS_Step2}
\alias{mcRigorTS_Step2}
\title{Step 2 of mcRigor two-step: Re-partition single cells that previously formed dubious metacells into smaller metacells.}
\usage{
mcRigorTS_Step2(
  step1_res,
  obj_singlecell,
  cell_membership_twostep,
  twostep_gamma = NULL,
  method = c("seacells", "mc1", "mc2", "supercell", "metaq"),
  custom_thre = F,
  color_field = NULL
)
}
\arguments{
\item{step1_res}{A list of results obtained form Step 1 of mcRigor two-step, given by function mcRigorTS_Step1.}

\item{obj_singlecell}{A Seurat object of the original single cells.}

\item{cell_membership_twostep}{A dataframe, each column of which contains the metacell membership of single cells that need to be re-partitioned under a given gamma (granularity level). 
The column names should be the corresponding gamma's (in the character type). The row names should be the single cell id's.}

\item{method}{The metacell partitioning method used to build metacells. Available: "seacells" (default, SEACells), "mc1" (MetaCell), "mc2" (MetaCell2), "supercell" (SuperCell), "metaq" (MetaQ).}

\item{custom_thre}{A boolean indicating whether to re-compute the threshold values.}

\item{color_field}{The variable based on which to color the cells.}

\item{TabMC}{A dataframe containing the permutation results and mcRigor results in dubious metacell detection. Saved in the previous steps through function mcRigor_DETECT.}
}
\value{
A list containing the following fields: 

\item{obj_sc_dub}{A Seurat object of single cells that are marked as dubious and will be re-partitioned.}
\item{target_res}{A list containing the permutation results, derived thresholds, and mcRigor results in dubious metacell detection, given by function mcRigor_threshold.}
\item{obj_metacell_step1}{The metacell object built under the membership sc_membership, with more dubious metacells marked.}
}
\description{
Step 2 of mcRigor two-step: Re-partition single cells that previously formed dubious metacells into smaller metacells.
}
