\name{mcRigor-package}
\alias{mcRigor-package}
\alias{mcRigor}
\docType{package}
\title{
\packageTitle{mcRigor}
}
\description{
\packageDescription{mcRigor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mcRigor}
\packageIndices{mcRigor}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mcRigor}

Maintainer: \packageMaintainer{mcRigor}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
